Adds mend_dist process and R script, upd Docker
mend_dist process requires an R script to run. The script
mendelerror_family_plotting.R was added to bin folder. This script
requires a number of R packages that have not been previously instaled
to main siteqc docker container, thus the container was updated.

Some R packages were installed with conda, but not all. The
package stringi from anaconda cloud (both r and conda-forge repos)
had problems with installation to Debian 10. It complained about missing
libicu64 package files, and this package is not available for Debian 10.
The only solution I found was to install stringi (and all packages that
depend on it) with R itself. This is slower because it compiles packages
from source, but in this case R downloaded available icu61 for Debian 10
and compiled stringi from source with icu61.

Known issue:
The R script added for mend_dist process is not designed to work with
only one input chromosome chunk file - it needs at least two as I
understand. The issue comes in 1) line 27: `summary(fmendel$N)` where
fmendel should be an array of values, but when there is only one input
bcf file its not an array and R throws an error. And 2) ggplot needs at
least two data points for each layer of the graph, and as I understand it
complains that it didn't have enough with one file. But also it may be
because we have one family, not because we have one chunk file.