Fixes mend_dist R script to print summary() output
The issue was that summary() outputs a "summary" type file, that cannot
be coerced to a dataframe and thus can't be saved with write.csv().

I assume the only purpose of saving this file was to have a later look
at it by researcher, so it shouldn't matter how exactly it is saved, as
long as all info is retained and its human-readible. So I decided to fix
the error with capture.output(). No error and looks nicer.

Alongside with this commit I slightly changed the input files on s3:
- added a pair of files for chr11 that is just a renamed chr10;
- added family2 with 3 participants to example.fam and keep.txt files.
Changed input files now allow to run more than one instance of a process
that helps to find some caveats of pipeline, like missing "each", etc.