Adds aggregateAnnotations process
The step is not finished untill the R script is fixed.

From Nextflow side - this step requires a complex mapping of channels
to ensure that only files generated from same bcf input file of a
particular genomic region are aggregated in this step. For this reason
all channels have been supplemented with a "key" `region` value that was
used to join corresponding files together.