Fixes and adds processes, updates docker container (#9)
* Fixes 4 processes, updates docker container

- Three processes missed asort() function for awk, that is distrubuted
with GNU awk only. Solved by installing gawk to main siteqc container.
- Fourth process missed quotes around awk command and wrong output files
name.

Additionally:
- Bound docker container to inherit for nfcore/base with an immutable tag.

* Adds process mend_err_p1 with parameters

* Adds optional --keep file argument to mend_err_p1

For now the pipeline is set to take this file as general input of
pipeline, but in final version of pipeline it might be genreated in the
pipeline itself (as well as example.phe file) by triodata plink process.

* Adds process mend_err_p2 with needed example file

* Adds mend_dist process and R script, upd Docker

mend_dist process requires an R script to run. The script
mendelerror_family_plotting.R was added to bin folder. This script
requires a number of R packages that have not been previously instaled
to main siteqc docker container, thus the container was updated.

Some R packages were installed with conda, but not all. The
package stringi from anaconda cloud (both r and conda-forge repos)
had problems with installation to Debian 10. It complained about missing
libicu64 package files, and this package is not available for Debian 10.
The only solution I found was to install stringi (and all packages that
depend on it) with R itself. This is slower because it compiles packages
from source, but in this case R downloaded available icu61 for Debian 10
and compiled stringi from source with icu61.

Known issue:
The R script added for mend_dist process is not designed to work with
only one input chromosome chunk file - it needs at least two as I
understand. The issue comes in 1) line 27: `summary(fmendel$N)` where
fmendel should be an array of values, but when there is only one input
bcf file its not an array and R throws an error. And 2) ggplot needs at
least two data points for each layer of the graph, and as I understand it
complains that it didn't have enough with one file. But also it may be
because we have one family, not because we have one chunk file.

* Adds process pull_1kg_p3_sites (pulls from s3)

* Fixes mend_err_p2 to not consume predefined_fam

* Fixes mend_dist R script to print summary() output

The issue was that summary() outputs a "summary" type file, that cannot
be coerced to a dataframe and thus can't be saved with write.csv().

I assume the only purpose of saving this file was to have a later look
at it by researcher, so it shouldn't matter how exactly it is saved, as
long as all info is retained and its human-readible. So I decided to fix
the error with capture.output(). No error and looks nicer.

Alongside with this commit I slightly changed the input files on s3:
- added a pair of files for chr11 that is just a renamed chr10;
- added family2 with 3 participants to example.fam and keep.txt files.
Changed input files now allow to run more than one instance of a process
that helps to find some caveats of pipeline, like missing "each", etc.

* Add mend_err_p3 process

* Adds aggregateAnnotations process

The step is not finished untill the R script is fixed.

From Nextflow side - this step requires a complex mapping of channels
to ensure that only files generated from same bcf input file of a
particular genomic region are aggregated in this step. For this reason
all channels have been supplemented with a "key" `region` value that was
used to join corresponding files together.

* Fixes aggregate_annotaition R script and queries

R script had three issues:
- loading libratries - simple fix;
- fifth column in startfile - complex fix;
- king env var missing - simple fix.

main.nf - changed output folder and file to be coherent with script.

config - all querries were missing part to generate fifth column of ID.

King env var was outourced to config and by default is True. This is
needed to follow a particular execution pathway of agg_annot R script
that fits out input files at that stage of pipeline development.

All bcf querries had to be updated to cantaint fifth part of ID column.
Otherwise agg_annot R script failed. And this column has to be identical
for all files generated from bcf input, otherwise they can't be joined
afterwards and result file is empty.