PR to finish SiteQC pipeline: adds triodata, conditional rule execution, upgrades aggrAnnot (#12)
* Makes mend_dist and mend_err_p2 run conditionally

These processes don't have to be run if user provides pheno keep
file to rule mend_err_p3. (With --mend_err_p3_keep_pheno argument.
Example name is "MendelFamilies_4SD.fam").

* Adds updated trio_define.r from Daniel (#13)

Daniel uploaded the updated script to slack on 2nd Septemer and asked to use it instead of previous one.
The source thread: https://lifebit-biotech.slack.com/archives/C0180RJ6C07/p1599041508035300
@cgpu asked to upload the script with a separate PR to easily see the changes.

Comment from Daniel on changes made to the script:
"This is a modified version of script - there are some changes that should deal with package conflicts and a change in data padding"

Commits:
* adds_updated_trio_define.r_from_Daniel

Daniel uploaded updated script to slack on 2nd Septemer and asked to use
it instead of previous one.

* Fixes small copying error

* Adds triodata_define process, conditionally

This process is producing keep.txt and triodata.fam if not provided by
user. If both files are provided, step is skipped.

* Updates aggregate_annotation process

Updates the process to lates't version from Daniel, from 15th and
17th Sep messages. For more details look into PR #14 of current repo.

* Changes s3://.../nodata into params.empty_file

* Updates help message to include new args

* Adds '[file]' in help message where missing

* Update usage.md